My experiment failed, but I learned a lot at least.
My goal was to switch to C in real mode.

I was able to concat my nasm artifact and my cc artifact together
and actually run C code, which was great!

The problem is that the machine instructions emitted are still
using 32 bit registers, and I was getting crazy undefined behavior.

I tried setting up a toolchain to use gcc 5.5 in docker. That
was a rabbit hole of applying patches to get it to build, then
I saw this post:
https://stackoverflow.com/a/19055849

So even back then, it didn't work. There is also bruce's C compiler,
which DOES emit 16 bit insturctions, but that's going way off the beaten
path at that point. Best to just stick to assembly until if/when I switch to
protected mode.

I also learned a lot about objdump, linker scripts, and make, which is good for the future
, as well as how inline assembly in gcc works.

Revert "--wip-- [skip ci]"

This reverts commit 1b20c5a7111228f087e821e30a46c3848e68ba89.