ok, I'm done for now. what I've learned is that the emulated card in QEMu only emuluates correctly if you're in protected mode. Specifically, `RxBuf` in etl18139.c is in `rtl8139_io_writel` but not `rtl8139_io_writew` or `rtl8139_io_writeb`.
Switching to protected mode is a whole other can of worms that I was hoping to avoid and
makes this a lot more work.

The other thing to remember when I come back next time is the whole implicit ds / nasm [] address thing.

Also, the vga routine works, just very primitiive atm.